+++ heftia-effects/src/Control/Effect/Interpreter/Heftia/Concurrent/Pipe/IO.hs
@@ -2,7 +2,7 @@
 -- License, v. 2.0. If a copy of the MPL was not distributed with this
 -- file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-module Control.Effect.Handler.Heftia.Concurrent.Pipe.IO where
+module Control.Effect.Interpreter.Heftia.Concurrent.Pipe.IO where
 
 import Control.Arrow ((>>>))
 import Control.Effect (type (~>))