+++ heftia-effects/ChangeLog.md
@@ -8,3 +8,10 @@
 
 * Redesign from scratch.
 * Released as a beta version.
+
+## 0.3.0.0 -- 2024-09-01
+
+* Added interpreters for 'Timer' effects.
+* Simplify the logging example.
+* Update the data-effects version to 0.1.1.
+* Rename the module from `Control.Effect.Handler` to `Control.Effect.Interpreter` to align with terminology.